talita-nvidia-v1-gpu/requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-nvidia-ai-endpoints
langchain-community
pillow
faiss-gpu